TKT-68: ai-review fixes — bigint quarantine schema, drop tautological assertion
A schema value past int32 passes the consumer's future-variant check but overflowed
the integer column: the INSERT range error fell into the generic retry arm and looped
one malformed event as a permanent delayed NAK with readiness latched (finding 1).
The column is now bigint and a smoke test pins the huge-schema write. Also removes a
dead always-false clause from the msg-id determinism assertion (finding 2).

+++ services/inventory/internal/store/migrations/0008_catalog_event_quarantine.sql
@@ -8,7 +8,10 @@
 CREATE TABLE catalog_event_quarantine (
     subject text NOT NULL,
     event_id uuid NOT NULL,
-    schema integer NOT NULL CHECK (schema > 0),
+    -- bigint, not integer: the consumer forwards any envelope schema above its known max here,
+    -- and an int32-overflowing value must quarantine like any other future variant — an INSERT
+    -- range error would send one malformed event into a permanent NAK loop (ai-review finding 1).
+    schema bigint NOT NULL CHECK (schema > 0),
     envelope bytea NOT NULL,
     first_seen_at timestamptz NOT NULL DEFAULT now(),
     last_seen_at timestamptz NOT NULL DEFAULT now(),
